Fix bug in provided vim commands
OpenInVSCode and OpenCwdInVSCode would fail when the line number started
with "8".

This is because "%:813:4" (for example) evaluates the ":8" as a modifier to
the "%" (see help %:8) and thus the remainder of the number would get
appended to the filename and then the column param would get interpreted
as the line number param but wouldn't make any difference because the
accidentally-modified filename will open an empty document in vscode.

e.g. "%:813:4" --> "filepath13:4"